requirements.txt
langchain
langchain-openai
langchain-ollama
python-dotenv